Cast krb5_error_code to int to avoid warning.
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@22993 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/ticket.c
@@ -148,7 +148,7 @@ find_type_in_ad(krb5_context context,
 					   NULL);
 	    if (ret) {
 		krb5_set_error_string(context, "Failed to decode "
-				      "IF_RELEVANT with %d", ret);
+				      "IF_RELEVANT with %d", (int)ret);
 		goto out;
 	    }
 	    ret = find_type_in_ad(context, type, data, found, FALSE,
